refactorgameplay.jsx and gameplay.css
