+++ time-react/frontend/src/components/Header/Header.jsx
@@ -1,11 +1,11 @@
 import './Header.css'
-import logo from '../../assets/logo.svg';
+import logo from '../../assets/logo.png';
 
 function Header() {
     return (
         <header>
             <div id="logo">
-                <img src={logo} alt="TimeLink-game logo"/>
+                <img src={logo} alt="TimeLink-game logo" />
             </div>
             <h1>TimeLink</h1>
         </header>
